agrego servicios y repositorios

+++ src/main/java/FIUBA/CineXplore/model/Actor.java
@@ -18,7 +18,7 @@ import java.util.Set;
 public class Actor {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer actorId;
+    private Long actorId;
 
     @Column(length = 150, nullable = false)
     private String fullName;
